requirements.txt
anytree
flask
jsonpickle
lark-parser
pymongo

pandas==0.23.0
numpy==1.14.3
scipy==1.1.0
scikit-learn==0.19.1